+++ changelog.md
@@ -4,6 +4,7 @@ Nav Beacon now costs Aluminium instead of Iron<br>
 Helium-3 Mine now costs Aluminium instead of Copper<br>
 Elerium Mining Ship now costs Titanium instead of Iridium<br>
 Iron Mining Ship now costs Aluminium instead of Titanium<br>
+Dimension Compression upgrade now works correctly<br>
 
 #### 0.4.13 - 6/26/2019 ####
 Aluminium resource added to the game, research Bayer Process to unlock Metal Refinery which in turn unlocks Aluminum<br>